% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_figure_1.R
\name{Make_figure_1}
\alias{Make_figure_1}
\title{Extracting the first plot}
\usage{
Make_figure_1(x)
}
\arguments{
\item{x:an}{expression matrix}
}
\value{
the boxplot
}
\description{
Extracting the first plot
}
